# Improve startup performance