Several changes targeted specifically at avoiding allocations during startup. By allocating less we make less work for the GC to clean up after us, which accumulates in saved time during loading from fewer GC collections. See commit messages for motivation for each change.

Startup time improvements for the default mods on my machine:

| Mod | Bleed  | PR |
| - | - | - |
| d2k | 5.22s | 3.77s |
| ts    | 6.50s | 6.11s |
| ra    | 5.79s | 5.56s |
| cnc  | 3.91s | 3.82s |

Each mod run several times until it produced consistent timing results (the first couple of runs tend to be slower until files are in the OS cache) then took the average of 6 runs. Timing given by making the following change in [Game.cs](https://github.com/OpenRA/OpenRA/blob/14fc0254c69ab376dd424ed6efb5f11705b017f2/OpenRA.Game/Game.cs#L268-L275):

```diff
public static RunStatus InitializeAndRun(string[] args)
{
	Initialize(new Arguments(args));
 
	// Proactively collect memory during loading to reduce peak memory.
	GC.Collect();
-	return Run();
+	Console.WriteLine(stopwatch.Elapsed.TotalSeconds.ToString("0.00s"));
+	return RunStatus.Success;
}
```