@pchote 2020-10-18
I find it interesting that D2k goes from taking significantly longer than TD to slightly faster. Is this because the R8 parser was leaning more heavily on the garbage creating stream reads?