@RoosterDragon 2020-10-20
I can't be sure exactly why the impact in d2k is so noticeable, but the R* speculation is probably fair.